Extract `assetCompareFunction` to new function
